image-processing/requirements.txt
flask
python-dotenv
psycopg2-binary
flask-sqlalchemy
flask-migrate
pyjwt
pika
Pillow
gunicorn